--- app/models.go
@@ -1,12 +1,0 @@
-package app
-
-import (
-	"go-rpg/proto"
-
-	"gorm.io/gorm"
-)
-
-type character struct {
-	gorm.Model
-	*proto.Character
-}
