+++ db/db.go
@@ -3,6 +3,7 @@ package db
 import (
 	"os"
 
+	"github.com/glebarez/sqlite"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
@@ -11,5 +12,12 @@ func Postgres() (*gorm.DB, error) {
 	dsn := os.Getenv("POSTGRES_CONN")
 	return gorm.Open(postgres.Open(dsn), &gorm.Config{
 		DisableForeignKeyConstraintWhenMigrating: true,
+	})
+}
+
+func InMemory() (*gorm.DB, error) {
+	return gorm.Open(sqlite.Open(":memory:"), &gorm.Config{
+		DisableForeignKeyConstraintWhenMigrating: true,
+		PrepareStmt:                              true,
 	})
 }
